added custom output symbols
Instead of OK / FAILED(n), output symbols can be customized. Per default, it is still Ok / Failed(n). A user has now the possibility to define a custom output if the execution of an input symbol succeeds and he can define a specific output for the failure of an action.