Started the next, big refactoring.
The frontend got moved into its very own module.
The structure of the backend got shuffled around a lot. Instead of only organizing the classes by function, they are now organized by their logical unit and function. This revealed that in the backend a lot of code is connected to many different areas, which still needs fixing.
The backend and frontend are currently build separately, which means that ALEX is not so easy to run. This is the next thing to fix.

Please also note: The tests are broken and this commit does not fix them.