added test case steps
A test case step either executes a symbol or a single action. This allows a user to add logic between symbols that would usually not fit in the modelling of a symbol, but is required for a test case. 

Further, this way, a test case can be modelled independently from symbols.