More changes to make Travis happy.
Changed the log level of org.glassfish from 'info' to 'warn' to further
reduce the log spam.
Travis now always runs with the 'bootstrap' profile enabled. This fixes
problems with the node_modules in the cache, e.g. this ensures that
always the newest versions are used during the build.