don't close the browser in each reset
Closing and opening the web browser window for each membership query is very costly. The browser instance is now reused during the whole learning process and closed at the end.

Since this kind of "soft reset" does not delete the browser cache or the history (but cookies, sessionstorage and localstorage), the action BrowserAction.java can be used to restart the browser, e.g. in the reset symbol.

Connectors now have to implement a post() method. Contrary to dispose(), which is called after each membership query, post() should only executed after a learning / testing process finished.