follow HTTP redirects manually
When making a request to a REST endpoint of the system under testing (SUT), it might answer with status 302. Jersey automatically redirects the requests to the new location. However, the HTTP response of the initial request might have important data, especially cookies, that we want to keep but get lost during the redirect.

With this commit, we disabled the automatic redirects and do it ourselves to keep the cookies of the initial request.