use parameterized symbols for learning
One can now make use of symbol parameters for learning, too. This way, a symbol can also be used multiple times in a learning process if the values for their parameters differ. The inputs in the model will be "symbol <val1, val2>" where "val1" und "val2" are the concrete values for the two parameters of the symbol. This also makes it easier to interop symbols between learning and testing, where symbol parameters have been used already.