make ChromeDriver path configurable via variable
Previously, the chromedriver path for running the testsuite-generator-test was hardcoded to '/usr/bin/chromedriver' and the test was skipped, depending on whether the file existed or not.

With this commit, it is possible to specify an alternative path, by setting the environment variable ALEX_CHROMEDRIVER_PATH. This allows to configure CI environments (or any other system with non canonical installation paths) to run the tests as well.