share symbolMapper instance across forked SULs
Previously, when forking a SUL, a new symbolMapper instance was created. This resulted in synchronization issues when adding new symbols to the learning process, because only the global mapper was updated and previously forked (mapped) SULs missed this update. This commits shares the same instance across multiple forks, which should not introduce any race-conditions, because the mapper is only accessed read-only _after_ the symbol has been added.